requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
libcst==0.3.23
lxml==4.7.1
multitasking==0.0.10
mypy-extensions==0.4.3
numpy==1.22.0
packaging==21.3
pandas==1.3.5
praw==7.5.0
prawcore==2.3.0
proto-plus==1.19.8
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.0
rsa==4.8
six==1.16.0
typing-inspect==0.7.1
typing_extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.7
websocket-client==1.2.3
yfinance==0.1.68
